requirements.txt
flask==2.3.2
pandas==1.5.3
scikit-surprise==1.1.3
flask-sqlalchemy==3.0.3